refactor use case create bank account
