+++ e2e/src/domains/account/application/use-cases/flows/fill-account-form.use-case.ts
@@ -1,0 +1,17 @@
+import { AccountPort } from '../../../domain/ports/ui/create-account-ui.port';
+import { CreateBankAccountRequest } from './create-bank-account-flow.use-case';
+
+export class FillAccountFormUseCase {
+  constructor(private readonly accountPort: AccountPort) {}
+
+  async execute(request: CreateBankAccountRequest): Promise<void> {
+    const formData = {
+      accountName: request.accountName,
+      accountType: request.accountType,
+      initialBalance: request.initialBalance,
+      currency: request.currency,
+      description: request.description,
+    };
+    await this.accountPort.fillAccountForm(formData);
+  }
+}
