+++ e2e/src/domains/account/application/use-cases/flows/navigate-to-form.use-case.ts
@@ -1,0 +1,10 @@
+import { AccountPort } from '../../../domain/ports/ui/create-account-ui.port';
+
+export class NavigateToFormUseCase {
+  constructor(private readonly accountPort: AccountPort) {}
+
+  async execute(): Promise<void> {
+    await this.accountPort.navigateToApp();
+    await this.accountPort.clickButton('Add Account');
+  }
+}
