+++ e2e/src/domains/account/application/use-cases/flows/submit-account-form.use-case.ts
@@ -1,0 +1,9 @@
+import { AccountPort } from '../../../domain/ports/ui/create-account-ui.port';
+
+export class SubmitAccountFormUseCase {
+  constructor(private readonly accountPort: AccountPort) {}
+
+  async execute(): Promise<void> {
+    await this.accountPort.submitForm();
+  }
+}
